+++ R/available-packages.R
@@ -234,6 +234,21 @@ renv_available_packages_entry <- function(package,
     entries[ordered[[1]], ]
   }
 
+  # if 'prefer' was supplied as a repository URL, include it
+  repos <- repos %||% as.list(getOption("repos"))
+  if (is.character(prefer)) {
+
+    # ensure repositories are named
+    nms <- names(prefer) %||% rep.int("", length(prefer))
+    nms[!nzchar(nms)] <- prefer[!nzchar(nms)]
+    names(prefer) <- nms
+
+    # include any 'prefer' repositories that were supplied as URLs
+    isurl <- grep("^\\w+://", prefer)
+    repos <- c(prefer[isurl], repos)
+
+  }
+
   # read available packages
   dbs <- available_packages(
     type  = type,
@@ -242,8 +257,8 @@ renv_available_packages_entry <- function(package,
   )
 
   # if a preferred repository is marked and available, prefer using that
-  if (length(prefer) == 1L && prefer %in% names(dbs)) {
-    idx <- match(prefer, names(dbs))
+  if (is.character(prefer)) {
+    idx <- omit_if(match(names(prefer), names(dbs)), is.na)
     ord <- c(idx, setdiff(seq_along(dbs), idx))
     dbs <- dbs[ord]
   }
@@ -280,10 +295,8 @@ renv_available_packages_entry <- function(package,
 
 renv_available_packages_record <- function(entry, type) {
 
-  # check to see if this is already a proper record
-  attrs <- attributes(entry)
-  keys <- c("type", "url")
-  if (all(keys %in% names(attrs)))
+  # check if this record was already tagged
+  if (renv_record_tagged(entry))
     return(entry)
 
   # otherwise, construct it
@@ -302,13 +315,19 @@ renv_available_packages_record <- function(entry, type) {
   # form url
   url <- entry$Repository
   path <- entry$Path
+  name <- entry$Name
+
+  # if path was supplied, it should be used relative
+  # to the repository URL
   if (length(path) && !is.na(path))
     url <- paste(url, path, sep = "/")
 
-  attr(record, "type") <- type
-  attr(record, "url")  <- url
-
-  record
+  renv_record_tag(
+    record = record,
+    type   = type,
+    url    = url,
+    name   = name
+  )
 
 }
 
@@ -484,10 +503,12 @@ renv_available_packages_latest_mran <- function(package,
   url <- file.path(base, name)
 
   # tag record with url + type
-  attr(record, "url")  <- dirname(url)
-  attr(record, "type") <- "binary"
-
-  record
+  renv_record_tag(
+    record = record,
+    type   = type,
+    url    = dirname(url),
+    name = "RSPM"
+  )
 }
 
 renv_available_packages_latest_repos <- function(package,