+++ tests/testthat/test-repos.R
@@ -59,10 +59,19 @@ test_that("repository names are always set", {
   skip_on_cran()
   renv_tests_scope()
 
-  repos <- c(CRAN = "1", "2", "3")
+  repos <- c(
+    CRAN = "https://cran.r-project.org",
+    PPM  = "https://packagemanager.posit.co",
+    "https://renv.github.io/drat"
+  )
+
+  expected <- c(
+    CRAN = "https://cran.r-project.org",
+    PPM  = "https://packagemanager.posit.co",
+    "https://renv.github.io/drat" = "https://renv.github.io/drat"
+  )
 
   actual <- renv_repos_normalize(repos)
-  expected <- c(CRAN = "1", V1 = "2", V2 = "3")
   expect_equal(actual, expected)
 
 })